feature: 도서 즐겨 찾기 기능 구현 및 BookReport 화면, BookDetail 화면 이동하지 않는 문제 수정

+++ core/database/src/main/java/com/twitty/database/dao/BookReportDao.kt
@@ -21,7 +21,7 @@ interface BookReportDao {
     suspend fun updateBookReport(bookReportEntity: BookReportEntity)
 
     @Query("SELECT * FROM book_reports WHERE id = :id LIMIT 1")
-    suspend fun fetchBookReport(id: Int): List<BookReportEntity>
+    suspend fun fetchBookReport(id: Long): List<BookReportEntity>
 
     @Query("SELECT * FROM book_reports")
     fun fetchBookReports(): Flow<List<BookReportEntity>>
