+++ core/database/src/main/java/com/twitty/database/model/BookReportEntity.kt
@@ -6,11 +6,11 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "book_reports")
 data class BookReportEntity(
     @PrimaryKey(autoGenerate = true)
-    val id: Int = 0,
+    val id: Long = 0,
     val title: String,
     val content: String,
     val date: String,
     val isFavorite: Boolean,
-    val bookId: Int,
+    val bookId: Long,
     val tagIds: List<Int>
 )
